project/requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
blindspin==2.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
docker==4.2.0
execnet==1.7.1
Faker==4.1.0
fasteners==0.15
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jenkins-job-builder==3.3.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.3.0
multi-key-dict==2.0.3
packaging==20.4
pbr==5.4.5
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pylint==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-jenkins==1.7.0
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.17
stevedore==1.32.0
testcontainers==3.0.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.9
vk-api==11.8.0
wcwidth==0.1.9
webdriver-manager==2.5.2
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
